Use float64 for outline rect thickness.

+++ drawer.go
@@ -56,7 +56,7 @@ type Drawer interface {
 		Whereas a value of OutlineInside will result in the outside edge of the outline
 		matching the rectangle coordinates.
 	*/
-	DrawOutlineRect(rect geometry.RectFloat, thickness float32, position OutlinePosition, colour color.Color)
+	DrawOutlineRect(rect geometry.RectFloat, thickness float64, position OutlinePosition, colour color.Color)
 
 	/*
 		Bell sounds a bell.
